requirements.txt

Django<1.9
django-appconf==1.0.1
django-autocomplete-light==3.1.1
django-celery==3.1.16
django-choices-enum==0.2
django-common-helpers==0.8.0
django-compressor==1.5
django-cors-headers==2.1.0
django-cron==0.4.6
django-enumfields==0.7.4
django-environ==0.4.4
django-extensions==1.9.6
django-filter==1.1.0
django-mathfilters==0.3.0
django-nose==1.4.5
django-redis-cache==1.6.2
django-redis-sessions==0.5.0
djangorestframework==3.3.3
djangorestframework-camel-case==0.2.0

requests==2.19.1
requests-mock==0.6.0
redis==2.10.3
psycopg2-binary==2.7.5
celery==3.1.18
google-auth==1.3.0
gunicorn==19.8.1
httplib2==0.9.2
Pillow==5.2.0
psutil==3.2.2
pyOpenSSL==18.0.0
python-dateutil==2.4.2
pytz==2018.5
PyYAML==3.11
uritemplate==0.6

colorlog==2.6.0
coverage==4.0
mock==2.0.0

bunch==1.0.1
enum34==1.0.4
fake-factory==0.6.0
futures==3.0.3
incf.countryutils==1.0

braintree==3.23.0
paypalrestsdk==1.11.5

WeasyPrint==0.42.3
Werkzeug==0.12.2

# autopep8==1.3.5
# flake8==3.5.0
# django-threadlocals==0.8 # is not used
# social-auth-app-django==2.1.0
# raven==6.6.0
